+++ modules/core/modelList/list.js
@@ -23,7 +23,6 @@ import Filter from "../../tools/filter/model";
 import PrintV2 from "../../tools/print/model";
 import Print from "../../tools/print_/mapfish3PlotService";
 import HighResolutionPrint from "../../tools/print_/highResolutionPlotService";
-import Download from "../../tools/download/model";
 import Animation from "../../tools/pendler/animation/model";
 import Lines from "../../tools/pendler/lines/model";
 import Contact from "../../tools/contact/model";
@@ -272,9 +271,6 @@ const ModelList = Backbone.Collection.extend(/** @lends ModelList.prototype */{
             }
             else if (attrs.id === "shadow") {
                 return new Shadow(attrs, options);
-            }
-            else if (attrs.id === "download") {
-                return new Download(attrs, options);
             }
             else if (attrs.id === "searchByCoord") {
                 return new SearchByCoord(attrs, options);