+++ src/modules/tools/stateTools.js
@@ -1,9 +1,9 @@
 import Draw from "./draw/components/Draw.vue";
-import ScaleSwitcher from "./scaleSwitcher/components/ScaleSwitcher.vue";
-import SupplyCoord from "./supplyCoord/components/SupplyCoord.vue";
 import FileImport from "./fileImport/components/FileImport.vue";
 import Gfi from "./gfi/components/Gfi.vue";
 import Measure from "./measure/components/Measure.vue";
+import ScaleSwitcher from "./scaleSwitcher/components/ScaleSwitcher.vue";
+import SupplyCoord from "./supplyCoord/components/SupplyCoord.vue";
 
 /**
  * User type definition
@@ -14,11 +14,11 @@ import Measure from "./measure/components/Measure.vue";
 const state = {
     componentMap: {
         draw: Draw,
-        scaleSwitcher: ScaleSwitcher,
-        supplyCoord: SupplyCoord,
         fileImport: FileImport,
         gfi: Gfi,
-        measure: Measure
+        measure: Measure,
+        scaleSwitcher: ScaleSwitcher,
+        supplyCoord: SupplyCoord
     },
     configuredTools: []
 };